+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Added `/iguanatweaks get_treasure_enchantments_book` command
+
 ## 3.14.5
 * Fixed `iguanatweaksreborn:reduced_drop_no_player_kill_blacklist` not loading when progressive bosses wasn't installed
 * Fixed mixin error when serene season was not installed